backend-app/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-tree-queries==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
cffi==1.15.1
cryptography==41.0.4
ecdsa==0.18.0
pyasn1==0.5.0
pycparser==2.21
python-jose==3.3.0
rsa==4.9
six==1.16.0
XlsxWriter==3.2.0
requests==2.32.3
